decl_mode: Add missing "." in decl_mode grammar.
This seemed to be causing failure with coqc, and intermittent failure with
interactive mode.  The reason it only caused intermittent failure in
interactive mode is that it would successfully parse and execute a command up
to a ".", and then fail. But if it is being passed commands one phrase at a
time, every command would still be executed.

Signed-off-by: Tom Prince <tom.prince@ualberta.net>