Add option to dump universes to a file when an inconsitency is detected.
This needs a bit of cleanup, dump_universes_gen should be moved to a module
that both Cerrors and Vernacentries can access.

Also, I am not sure about the option naming, and perhaps there should be
separate options to enable the dumping and to control the name of the file.

Signed-off-by: Tom Prince <tom.prince@ualberta.net>