simpl honors "Recursive Arguments" declarations
The problem that this _proof_of_concept_ patch tries to solve is that
sometimes the unfolding behaviour of simpl is too aggressive, frequently
exposing match constructs. Moreover one may want some constants to never
be unfolded by simpl (like in the Math-Comp library where the nosimpl
hack is pervasive). Finally, one may want some constants to be volatile,
always unfolded by simple when applied to enough arguments, even if they
do not hide a fixpoint or a match.

The new vernacular "Recursive Arguments" attaches to constants an
extra-logical piece of information. It declares which arguments must be
checked to evaluate to a constructor before performing the unfolding of
the constant (in the same spirit simpl considers as such the recursive
argument of an inner fixpoint). It also specifies the minimum number of
arguments to which the constant must be applied to be unfolded.

Example: after declaring that both the first and the second arguments of
minus have to be considered as recursive by "simpl"

  Recursive Arguments minus [1 2].

"simpl" behaves as follows:

  (S x - y)   --- simpl ---> (S x - y)

  (S x - S y) --- simpl ---> (x - y)

Since the "Recursive Arguments" info is attached to "minus" only

  (S x - y)   --- unfold minus --->
                       match y with 0 => S x | S y' -> x - y' end

That is the same result one obtains without this patch calling "simpl".

A constant like "function composition" does not hide a fixpoint but with
this patch can be unfolded by simpl if applied to enough arguments (6
actually), none of which needs to evaluate to a constructor.

  Definition fcomp A B C (f : B -> C) (g : A -> B) x := f (g x).
  Recursive Arguments fcomp [] 6.
  Implicit Arguments fcomp [A B C].
  Infix "\o" := fcomp (at level 50, left associativity) : nat_scope.

  (fun x => (f \o g) x) = f \o g  --- simpl ---> (fun x => f (g x)) = f \o g

To get non-simplifiable constants one can just specify a number of
arguments that will never be available (in a well typed term of course).

  Recursive Arguments minus [] 3.

  (S x - S y) --- simpl ---> (S x - S y)

Note that this non-reducing behaviour affects only "simpl", for the rest
of the system (S x - S y) and (x - y) are still convertible.

Last, volatile constants can be obtained as follows

  Definition nid (x : nat) := x.
  Recursive Arguments nid [].

  nid   --- simpl ---> (fun x => x)

Note that the last integer argument of "Recursive Arguments" if omitted
is computed as the max of the integers between [], thus in the last case
it is 0 (nid applied to 0 or more arguments is unfolded), in the case of
minus it was computed to be two (both of which required to be concrete).

If there is interest in this feature, we may discuss this patch and the
required fixes at the next Coq GT.

Cheers