Add vernac command modifier to catch only anomalies.
bugs/closed/shouldsucceed/2304.v had a command that at one time raised an
anomaly. This was fixed so that it raised an error instead. However, the test
case started failing again, since the command no longer even gives an error.

This patch allows one to express that a command shouldn't raise an anomaly,
without caring whether it succeeds, or fails with a non-anomalous error.

Signed-off-by: Tom Prince <tom.prince@ualberta.net>